+++ HitThatLine/src/HitThatLine/Domain/Discussion/Post.cs
@@ -9,9 +9,9 @@ namespace HitThatLine.Domain.Discussion
         public string Id { get; set; }
         public string MarkdownText { get; private set; }
         public string DisplayText { get; private set; }
-        public long UpVotes { get; private set; }
-        public long DownVotes { get; private set; }
-        public long NetVotes { get { return UpVotes - DownVotes; } }
+        public int UpVotes { get; private set; }
+        public int DownVotes { get; private set; }
+        public int NetVotes { get { return UpVotes - DownVotes; } }
         public double Score { get; private set; }
         public DateTime CreatedOn { get; private set; }
         public DateTime LastActivity { get; private set; }
